CPR-305 try having a single base class for both web and queue integration tests
