+++ src/test/kotlin/uk/gov/justice/digital/hmpps/personrecord/controller/QueueAdminIntTest.kt
@@ -1,10 +1,9 @@
 package uk.gov.justice.digital.hmpps.personrecord.controller
 
 import org.junit.jupiter.api.Test
-import org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put
-import uk.gov.justice.digital.hmpps.personrecord.integration.IntegrationTestBase
+import uk.gov.justice.digital.hmpps.personrecord.integration.WebTestBase
 
-class QueueAdminIntTest() : IntegrationTestBase() {
+class QueueAdminIntTest() : WebTestBase() {
 
   @Test
   fun `should return HTTP Location header containing the URL of new person`() {
