+++ src/test/kotlin/uk/gov/justice/digital/hmpps/personrecord/seeding/PopulateFromNomisIntTest.kt
@@ -7,14 +7,14 @@ import org.assertj.core.api.Assertions.assertThat
 import org.awaitility.kotlin.await
 import org.awaitility.kotlin.untilAsserted
 import org.junit.jupiter.api.Test
-import uk.gov.justice.digital.hmpps.personrecord.integration.IntegrationTestBase
+import uk.gov.justice.digital.hmpps.personrecord.integration.WebTestBase
 import uk.gov.justice.digital.hmpps.personrecord.model.identifiers.CROIdentifier
 import uk.gov.justice.digital.hmpps.personrecord.model.identifiers.PNCIdentifier
 import uk.gov.justice.digital.hmpps.personrecord.model.types.SourceSystemType.NOMIS
 import java.time.LocalDate
 import java.util.concurrent.TimeUnit.SECONDS
 
-class PopulateFromNomisIntTest : IntegrationTestBase() {
+class PopulateFromNomisIntTest : WebTestBase() {
 
   @Test
   fun `populate from nomis`() {
